refactor: added one file per command
