+++ app/commands/exit.js
@@ -1,0 +1,14 @@
+import {sleep} from "../utils.js";
+
+const exit = {
+  id: "exit",
+  args: 0,
+  description: 'exit current session',
+  async exec(term, _args) {
+    term.writeln('terminating session...');
+    await sleep(1000);
+    window.location.href = 'https://www.youtube.com/watch?v=iik25wqIuFo';
+  },
+};
+
+export default exit;
